Re-Did Kene101 Power Module
I rebuilt kene101’s power module. However, the client_detail.php code
in the WIP was much different than the code of the current version.
This code works with the current 2.2.0 - but not haven’t been able to
test on the WIP. All that was throwing me is that I don’t see the
bluetooth tab-pane anymore.