Make the SIP check more resilient
The output of csrutil can be more than just one line of enabled or
disabled. If it's in a custom config, it will have '"enabled (Custom
Configuration)' in it, and many of the functions could still turned off.

Searching for just 'enabled' would make SIP seem active if just any of
the sup parts of SIP were enabled.