Drop Down the Login Page (#1062)
The text that pops over the lock icon at the login page is cutoff because the login box is too high up on the page. This PR pushes the login box down a bit so the text is visible. 

There maybe a more appropriate fix than a bunch of `<br>` tags but this works and looks good on a 27" iMac, 15" MBP and iPhone X.